+++ packages/frontend-airdos/src/pages/groups/GroupPage/GroupPageRouter.tsx
@@ -1,11 +1,9 @@
 import React from "react"
 import { Route, Redirect } from "react-router-dom"
 
-import type { Group } from "../../typings/api"
+import type { Group } from "../../../typings/api"
 
-const GroupPageContentRouter: React.FC<{ group: Group }> = ({
-  group: { name },
-}) => (
+const GroupPageRouter: React.FC<{ group: Group }> = ({ group: { name } }) => (
   <>
     <Route exact path={`/groups/${name}/`}>
       <Redirect to={`/groups/${name}/feed`} />
@@ -25,4 +23,4 @@ const GroupPageContentRouter: React.FC<{ group: Group }> = ({
   </>
 )
 
-export { GroupPageContentRouter }
+export { GroupPageRouter }
