+++ packages/frontend-airdos/src/pages/groups/GroupsOverviewPage/index.tsx
@@ -1,0 +1,1 @@
+export { GroupsOverviewPage } from "./GroupsOverviewPage"
