+++ packages/frontend-airdos/src/pages/groups/index.tsx
@@ -1,0 +1,1 @@
+export { GroupsRouter } from "./GroupsRouter"
